New dep-update inte-test: factor out a mixin
Put the _do_update in a single place, and have all of the test
classes import it.
Also, this adds tests for the mixed_operations one.
