+++ tests/integration_tests/tests/agentless_tests/deployment_update/__init__.py
@@ -168,3 +168,17 @@ class DeploymentUpdateBase(AgentlessTestCase):
     @staticmethod
     def _assertDictContainsSubset(subset, containing_dict):
         assert subset.items() <= containing_dict.items()
+
+
+class NewDeploymentUpdateMixin(object):
+    _group_update_workflow = 'csys_new_deployment_update'
+    _workflow_name = 'csys_new_deployment_update'
+
+    def _do_update(self, deployment_id, blueprint_id=None, **kwargs):
+        params = {
+            'blueprint_id': blueprint_id,
+        }
+        params.update(kwargs)
+        exc = self.client.executions.start(
+            deployment_id, 'csys_new_deployment_update', parameters=params)
+        self.wait_for_execution_to_end(exc)
