+++ tests/integration_tests/tests/agentless_tests/deployment_update/test_deployment_update_addition.py
@@ -21,7 +21,7 @@ from pytest import mark
 
 from integration_tests.tests.utils import (tar_blueprint,
                                            wait_for_blueprint_upload)
-from . import DeploymentUpdateBase, BLUEPRINT_ID
+from . import DeploymentUpdateBase, BLUEPRINT_ID, NewDeploymentUpdateMixin
 
 
 pytestmark = pytest.mark.group_deployments
@@ -561,14 +561,8 @@ class TestDeploymentUpdateAddition(DeploymentUpdateBase):
         self.assertRegexpMatches(deployment['description'], 'new description')
 
 
-class NewTestDeploymentUpdateAddition(TestDeploymentUpdateAddition):
-    def _do_update(self, deployment_id, blueprint_id=None,
-                   preview=False, **kwargs):
-        params = {
-            'blueprint_id': blueprint_id,
-        }
-        if preview:
-            params['preview'] = preview
-        exc = self.client.executions.start(
-            deployment_id, 'csys_new_deployment_update', parameters=params)
-        self.wait_for_execution_to_end(exc)
+class NewTestDeploymentUpdateAddition(
+    NewDeploymentUpdateMixin,
+    TestDeploymentUpdateAddition
+):
+    pass
